Fix thermal fluctuator
In previous versions, randDouble was allowed to generate a value of 0, which
would subsequently be passed as the argument to the log function. This, of
course, is disastrous. So, the range of randDouble is now (low, high).

Also, I modified network.cpp to throw an exception if NaN is ever assigned to
the position of a node, and integrator.cpp produces a little more debugging
information.