Fixed nonaffinity calculation
Before this version, there were problems with the calculation of the
nonaffinity parameter in the affine position/velocity calculations. Those have
been corrected. Now, the functions nonAffinity and nonAffinity_dd output the
squared displacement of the network and the squared relative velocities of the
network, accordingly. Therefore, to achieve the actual nonaffinity measure,
you must divide out by the strain/strain rate and the network size. However,
these are both preserved in the nonaffinity output file.