Fixed for hopefully the last time.
It looks like things are finally working. I think that the "roundoff error"
issues I've had in the past are really problems with my algorithm in the
xshift function. Also, the nonaffinity functions were wrong after all.
Hopefully everything is working as it should.

Added config files for the different sims

Fixed fluid velocity and nonaffinity_dd.

Now the fluid velocity and nonaffinity_dd should be correct.

Changed everything to numerical simulation

Now, the strain at each time step and the comparison to the nonaffine
case is calculated numerically. This should fix everything!

Use the integrated rather than theoretical strain

Fixed the printer module so that it uses the integrated strain
rather than the theoretical one.